1, fixed unused introducer. 2, Unified schema version number to int64_t

+++ src/meta/NebulaSchemaProvider.cpp
@@ -10,7 +10,7 @@
 namespace nebula {
 namespace meta {
 
-int32_t NebulaSchemaProvider::getVersion() const noexcept {
+int64_t NebulaSchemaProvider::getVersion() const noexcept {
     return ver_;
 }
